Splits accounting settings to individual components
